# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KDF NUTS eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, rating]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [admin]
      summary: Update category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      parameters:
        - in: query
          name: platform
          schema:
            type: string
            enum: [mobile, website, both]
          required: false
          description: Filter banners by platform
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [admin]
      summary: Create banner (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "201":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [admin]
      summary: Update banner (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "200":
          description: Banner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [admin]
      summary: Delete banner (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageUploadRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUploadResponse"
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve storage object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: string
        hex:
          type: string
        price:
          type: string
        stock:
          type: integer
        sku:
          type: string
      required: [id, name, value, stock]
    Product:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: string
        originalPrice:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        gradient:
          type: string
        tags:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        weight:
          type: string
        unit:
          type: string
        active:
          type: boolean
        featured:
          type: boolean
        rating:
          type: string
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, price, stock, active, featured, reviewCount]
    CreateProductRequest:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: string
        originalPrice:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        gradient:
          type: string
        tags:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        weight:
          type: string
        unit:
          type: string
        active:
          type: boolean
        featured:
          type: boolean
      required: [name, price, stock]
    ProductsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        altText:
          type: string
        color:
          type: string
        parentId:
          type: integer
        sortOrder:
          type: integer
        metaTitle:
          type: string
        metaDescription:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, active, sortOrder]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        altText:
          type: string
        color:
          type: string
        parentId:
          type: integer
        sortOrder:
          type: integer
        metaTitle:
          type: string
        metaDescription:
          type: string
        active:
          type: boolean
      required: [name, slug]
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        mobileImageUrl:
          type: string
        linkUrl:
          type: string
        targetType:
          type: string
          enum: [product, category, page]
        targetId:
          type: integer
        bgColor:
          type: string
        textColor:
          type: string
        label:
          type: string
        cta:
          type: string
        platform:
          type: string
          enum: [mobile, website, both]
        sortOrder:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, active, sortOrder]
    CreateBannerRequest:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        mobileImageUrl:
          type: string
        linkUrl:
          type: string
        targetType:
          type: string
          enum: [product, category, page]
        targetId:
          type: integer
        bgColor:
          type: string
        textColor:
          type: string
        label:
          type: string
        cta:
          type: string
        platform:
          type: string
          enum: [mobile, website, both]
        sortOrder:
          type: integer
        active:
          type: boolean
      required: [title]
...
    StorageUploadRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: number
        contentType:
          type: string
      required: [name, size, contentType]
    StorageUploadResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: number
            contentType:
              type: string
          required: [name, size, contentType]
      required: [uploadURL, objectPath, metadata]
...